requirements.txt
# project requirements
arrow==0.15.6
backcall==0.1.0
binaryornot==0.4.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
cycler==0.10.0
decorator==4.4.2
descartes==1.1.0
GDAL==3.0.4
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
Jinja2==2.11.2
jinja2-time==0.2.0
joblib==0.14.1
kiwisolver==1.1.0
laspy==1.7.0
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.1
packaging==20.1
pandas==1.0.3
parso==0.6.2
PDAL==2.2.1
pexpect==4.8.0
pickleshare==0.7.5
poyo==0.5.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.6
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtWebEngine==5.12.1
pyshp==2.1.0
python-dateutil==2.8.1
python-slugify==4.0.0
pytz==2019.3
requests==2.23.0
scikit-learn==0.22.1
scipy==1.4.1
Shapely==1.7.0
six==1.14.0
text-unidecode==1.3
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9

# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
